multi_agent_customer_support/pyproject.toml
[project]
name = "multi-agent-customer-support"
version = "0.1.0"
description = "Multi-agent customer support app using Google's ADK, FastAPI, and Supabase."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "google-adk",
  "a2a-sdk",
  "mcp",
  "a2a-mcp-server==0.1.5",
  "fastmcp==3.2.4",
  "fastapi",
  "uvicorn",
  "supabase",
  "python-dotenv",
  "httpx",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "pytest-html>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
